Un-nest DelegatingComponent
Move `DelegatingComponent` out from being a nested type.
